http2 v0.0.1 HTTP2.Huffman

Functions

decode(input)

Specs

decode(binary) :: binary

Decodes the given input using the Huffman code described at https://httpwg.github.io/specs/rfc7541.html#huffman.code.

Examples

iex> HTTP2.Huffman.decode(<<241, 227, 194, 229, 242, 58, 107, 160, 171, 144, 244, 255>>)
"www.example.com"

Input bytes that cannot be decoded are considered to be padding. Padding larger than 7 bits results in a decoding error:

iex> HTTP2.Huffman.decode(<< 511 :: 9 >>)
{:decoding_error, "Padding is larger than 7 bits"}

In addition padding bits have to match the most significant EOS (end-of-string) bits. Padding bits that don’t match result in a decoding error:

iex> HTTP2.Huffman.decode(<< 2 :: 6 >>)
{:decoding_error, "Padding does not match most significant bits of EOS"}

encode(input)

Specs

encode(binary) :: binary

Encodes the given input using the Huffman code described at https://httpwg.github.io/specs/rfc7541.html#huffman.code.

Examples

iex> HTTP2.Huffman.encode("www.example.com")
<<241, 227, 194, 229, 242, 58, 107, 160, 171, 144, 244, 255>>